+++ src/beamlime/dashboard/assemblers.py
